+++ CHANGELOG.md
@@ -1,6 +1,26 @@
 # CHANGELOG: Freeboard-SK
 
 ### v1.0.2
+### v1.1.0
+
+- Initial support for History and Playback APIs
+- Display Navigation data from `navigation.courseGreatCircle.nextPoint` when available
+- `Navigate To`: Select waypoint as course destination
+- Display `Prev / Next` point buttons to cycle through points of `Active Route`
+- Login / Authentication: Initial support.
+- Added `Zoom In / Out` buttons to map display
+- Update: Heading, TWD and AWA vessel lines to have consistant appearance at all scales
+- Update: Bearing line is now displayed only when course destination point data is available
+- Bug Fix: TWD & AWA lines not updated when SOG=0
+
+### v1.0.2
+
+- Fix: platform checks that were causing MS Edge to throw an exception.
+
+### v1.0.1
+
+- Fix: Url used to retrieve app list from non-node servers to include api version number.
+>>>>>>> r1.1
 
 - Fix: platform checks that were causing MS Edge to throw an exception.
 
@@ -38,5 +58,10 @@ To be able to set an anchor alarm the Signal K server will need to action HTTP P
 
 *If notification messages are received for the path `notifications.navigation.anchor` an alarm message will be displayed on the screen regardless of where the alarm was configured.*
 
+**Active Route / Navigate To:**
 
+To be able to set the Active Route or target Waypoint Signal K server will need to action HTTP PUT requests to the foloowing paths:
+1. `/signalk/v1/api/vessels/self/navigation/courseGreatCircle/nextPoint.position` 
+2. `/signalk/v1/api/vessels/self/navigation/courseGreatCircle/activeRoute.href` 
+3. `/signalk/v1/api/vessels/self/navigation/courseGreatCircle/activeRoute.startTime`